requirements.txt
alembic==1.3.3
aniso8601==8.0.0
Babel==2.8.0
Click==7.0
ecdsa==0.15
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.1
MarkupSafe==1.1.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycryptodome==3.3.1
python-dateutil==2.6.0
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==0.16.0
